flask-mvc/requirements.txt
python
flask
flask_cors
facenet_pytorch
numpy
pymongo
PyJWT
opencv-python
